% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4blog.R
\name{change2Tistoryhtml}
\alias{change2Tistoryhtml}
\title{Extract and change HTML to Tistory input html}
\usage{
change2Tistoryhtml(fileName, className = "language-r")
}
\arguments{
\item{fileName}{A name of your html file}

\item{className}{A className that your file to be}
}
\value{
html file but diffent className for the r code chunk
}
\description{
Extract and change HTML to Tistory input html
}
\examples{
# change2Tistoryhtml("abc.html")
# change2Tistoryhtml("abc.html", "language-r")
}
