% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4blog.R
\name{change_codeclass}
\alias{change_codeclass}
\title{Change code className in html file}
\usage{
change_codeclass(fileName, className)
}
\arguments{
\item{fileName}{A name of your html file}

\item{className}{A className that your file to be}
}
\value{
html file but diffent className for the r code chunk
}
\description{
Change code className in html file
}
\examples{
change_codeclass("abc.html", "language-r")
}
