% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4blog.R
\name{knitr4blog}
\alias{knitr4blog}
\title{knitr your Rmd file with className = "language-r"}
\usage{
knitr4blog(fileName, className = "language-r")
}
\arguments{
\item{fileName}{A name of your Rmd file}

\item{className}{A className that your R code class to be}
}
\value{
kniting your Rmd file
}
\description{
knitr your Rmd file with className = "language-r"
}
\examples{
# fileName should not have blank space
knitr4blog("abc.Rmd")
}
