% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4blog.R
\name{post2Tistory}
\alias{post2Tistory}
\title{Post your Rmd file to Tistory}
\usage{
post2Tistory(fileName, my_blogName, token, ...)
}
\arguments{
\item{fileName}{the .Rmd file name you want to post to your blog}

\item{my_blogName}{your blogName xxx from blog address "http://xxx.tistory.com"}

\item{token}{the token you have obtained from the url generated from token_url_maker function}

\item{...}{you can use className = "language-r" option in knitr4blog function}
}
\value{
posted blog on your blog
}
\description{
Post your Rmd file to Tistory
}
\examples{
# fileName should not have blank space
# Assume your file name is "test.Rmd"
# Assume your blog address is "http://issactoast.tistory.com"
my_token <- "your obtained token here"
# post2Tistory("test.Rmd", "issactoast", my_token)
}
