% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4blog.R
\name{token_url_maker}
\alias{token_url_maker}
\title{Tistory token-page url maker}
\usage{
token_url_maker(client_id, redirect_uri)
}
\arguments{
\item{client_id}{the id that is obtained from Tistory}

\item{redirect_uri}{the redirect uri that is submitted to Tistory}
}
\value{
url that ask tistory to generate token for blogging
}
\description{
Tistory token-page url maker
}
\examples{
my_id <- "your client_id here"
my_uri <- "your redirect_uri address here"
token_url_maker(my_id, my_uri)
}
