% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4blog.R
\name{token_url_maker}
\alias{token_url_maker}
\title{Tistory token-page url maker}
\usage{
token_url_maker(NoInput = NULL)
}
\arguments{
\item{NoInput}{this function has no inputs.}
}
\value{
browser automatically will be open containing url that token for
 posting. the redirected web page will be taining news of this package
}
\description{
Tistory token-page url maker
}
\examples{
# Getting token from url
# token_url_maker()
}
