% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4blog.R
\name{token_url_maker}
\alias{token_url_maker}
\title{Tistory token-page url maker}
\usage{
token_url_maker(client_id = "3a5ef7ff3d180eac94d5df5e58ba1768",
  redirect_uri = "http://issactoast.com/81")
}
\arguments{
\item{client_id}{the id that is obtained from Tistory}

\item{redirect_uri}{the redirect uri that is submitted to Tistory}
}
\value{
browser automatically will be open containing url that token for blogging
}
\description{
Tistory token-page url maker
}
\examples{
my_id <- "your client_id here"
my_uri <- "your redirect_uri address here"
token_url_maker(my_id, my_uri)
}
